Ability to pass custom options to handlers
This is very useful. For example your application's handlers could need the
URL of your OAuth provider and it could come from your application settings.

Now you can pass these options to the NsqSubscriber and they'll be passed over
to the handlers when a message arrives:

```ruby
subscriber = NsqSubscriber.new(
  lookupd: NSQ_LOOKUPD,
  topic: NSQ_TOPIC,
  channel: NSQ_CHANNEL,
  handler_options: {
    oauth_provider: OAUTH_PROVIDER,
    other_custom_stuff: "yes plese"
  }
)
```