Requeue messages on failure. Also improves backoff
The previous behaviour was to always confirm NSQ messages even when they failed
to be processed successfully.

Messages are requeued when an exception is raised. Messages are re-queue if the
number of attempts is under `:max_attempts` (15 by default). Discarded otherwise.

We also set the `:backoff_interval` (120s by default). This means the consumer
will slowdown when failures occur giving things the time to recover.

As far as I can tell from its code Krakow is using a linear backoff strategy [1].
It keeps track of the number of failures and it will use this number to determine
the backoff interval:

```ruby
registry_info[:backoff_until] = Time.now.to_i + (registry_info[:failures] * backoff_interval)
```

[1]: https://github.com/chrisroberts/krakow/blob/7d951699c9fc586b536b94b84e478925d9f6ca40/lib/krakow/distribution.rb#L203